Refactored LogSnapshot
