+++ lib/src/main/java/com/grio/lib/features/BugSnap.kt
@@ -8,7 +8,6 @@ import android.hardware.SensorManager
 import android.os.Bundle
 import android.util.Log
 import android.view.View
-import androidx.core.content.ContextCompat.startActivity
 import com.grio.lib.core.di.DaggerInjector
 import com.grio.lib.core.extension.screenshot
 import com.grio.lib.features.editor.DataHolder
@@ -52,7 +51,6 @@ class BugSnap {
             // Retrieve a SensorManager.
             val sensorManager = context.getSystemService(SENSOR_SERVICE) as SensorManager?
             var sd : ShakeDetector? = null
-            val log = LogSnapshot()
 
             // Register lifecycle.
             (context as Application).registerActivityLifecycleCallbacks(object: Application.ActivityLifecycleCallbacks {
@@ -69,7 +67,7 @@ class BugSnap {
                              activity?.window?.decorView?.post {
                                  val rootView = activity.window?.decorView as View
                                  val bitmap = rootView.screenshot()
-                                 val logDump = log.updateSnapshot()
+                                 val logDump = LogSnapshotManager.getLogTail()
 
                                  DataHolder.data = bitmap
